refactor: Move file Move Persons.java file to pojo package
